feat: standardize Suspense architecture across relationship and Strava components
- Convert RelationshipsList to AsyncRelationshipsList + RelationshipsListSkeleton pattern
- Convert RunnerSelector to AsyncRunnerSelector + RunnerSelectorSkeleton pattern
- Convert StravaDashboardWidget to AsyncStravaDashboardWidget + StravaDashboardWidgetSkeleton pattern
- Eliminate loadable utility usage in favor of direct async atom reads
- Create comprehensive skeleton components matching actual component structures
- Maintain consistent Mountain Peak design aesthetics across all components
- Improve perceived performance with proper loading states during Suspense
